CCM-12438: api level routing feature flag

+++ lambdas/backend-api/src/templates/app/routing-config-client.ts
@@ -40,6 +40,13 @@ export class RoutingConfigClient {
 
     if (clientConfigurationError) return clientConfigurationResult;
 
+    if (!clientConfiguration?.features.routing) {
+      return failure(
+        ErrorCase.VALIDATION_FAILED,
+        'Routing feature is disabled'
+      );
+    }
+
     if (!clientConfiguration?.campaignIds?.includes(validated.campaignId)) {
       return failure(
         ErrorCase.VALIDATION_FAILED,
@@ -73,6 +80,13 @@ export class RoutingConfigClient {
 
     if (clientConfigurationError) return clientConfigurationResult;
 
+    if (!clientConfiguration?.features.routing) {
+      return failure(
+        ErrorCase.VALIDATION_FAILED,
+        'Routing feature is disabled'
+      );
+    }
+
     if (!clientConfiguration?.campaignIds?.includes(validated.campaignId)) {
       return failure(
         ErrorCase.VALIDATION_FAILED,
@@ -91,6 +105,22 @@ export class RoutingConfigClient {
     routingConfigId: string,
     user: User
   ): Promise<Result<RoutingConfig>> {
+    const clientConfigurationResult = await this.clientConfigRepository.get(
+      user.clientId
+    );
+
+    const { data: clientConfiguration, error: clientConfigurationError } =
+      clientConfigurationResult;
+
+    if (clientConfigurationError) return clientConfigurationResult;
+
+    if (!clientConfiguration?.features.routing) {
+      return failure(
+        ErrorCase.VALIDATION_FAILED,
+        'Routing feature is disabled'
+      );
+    }
+
     return this.routingConfigRepository.submit(routingConfigId, user);
   }
 
@@ -98,6 +128,22 @@ export class RoutingConfigClient {
     routingConfigId: string,
     user: User
   ): Promise<Result<undefined>> {
+    const clientConfigurationResult = await this.clientConfigRepository.get(
+      user.clientId
+    );
+
+    const { data: clientConfiguration, error: clientConfigurationError } =
+      clientConfigurationResult;
+
+    if (clientConfigurationError) return clientConfigurationResult;
+
+    if (!clientConfiguration?.features.routing) {
+      return failure(
+        ErrorCase.VALIDATION_FAILED,
+        'Routing feature is disabled'
+      );
+    }
+
     const result = await this.routingConfigRepository.delete(
       routingConfigId,
       user
